refactor: Rename `prepareRelease` to `buildReleaseArtifacts`
- Move release artifact packaging into `app/build.gradle.kts`
- Rename the task to `buildReleaseArtifacts`
- Remove the separate `prepare-release.gradle.kts` script
